blog/requirements.txt
Django==1.11
gunicorn==19.9.0
django-markdownx
